Implemented locking.

+++ delete_recordings.pyw
@@ -1,10 +1,16 @@
 #!/usr/bin/env python
 
 import os
+import os.path
+
 import common
 
 def main():
-	common.delete_recordings()
-	
+	if common.get_lock("d"):
+		try:
+			common.delete_recordings()
+		finally:
+			common.release_lock()
+
 if __name__ == '__main__': 
 	main()
